mediasoup-types
New crate holding mediasoup types.

In order to be used by other crates to signal data to mediasoup crate and vice versa.

This crate permits not having to depend on the mediasoup crate which implies the C++
worker compilation.

This crate holds:
- data_structures
- rtp_parameters

The FBS conversion has been completely moved out of this type definition and it's
implemented in mediasoup data_structures.rs and rtp_parameters.rs via the corresponding
new FBS conversion traits defined in fbs.rs.

Closes #1498
