+++ Cargo.toml
@@ -2,5 +2,6 @@
 resolver = "2"
 members = [
     "rust",
+    "rust/types",
     "worker"
 ]
