+++ rust/tests/integration/consumer.rs
@@ -14,13 +14,13 @@ use mediasoup::rtp_parameters::{
     RtpHeaderExtensionDirection, RtpHeaderExtensionParameters, RtpHeaderExtensionUri,
     RtpParameters,
 };
-use mediasoup::scalability_modes::ScalabilityMode;
 use mediasoup::transport::ConsumeError;
 use mediasoup::webrtc_transport::{
     WebRtcTransport, WebRtcTransportListenInfos, WebRtcTransportOptions,
 };
 use mediasoup::worker::{Worker, WorkerSettings};
 use mediasoup::worker_manager::WorkerManager;
+use mediasoup_types::scalability_modes::ScalabilityMode;
 use parking_lot::Mutex;
 use std::net::{IpAddr, Ipv4Addr};
 use std::num::{NonZeroU32, NonZeroU8};
