+++ rust/types/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "mediasoup-types"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+once_cell = "1.16.0"
+serde_json = "1.0.87"
+thiserror = "2.0.12"
+
+[dependencies.serde]
+features = ["derive"]
+version = "1.0.190"
+
+[dependencies.regex]
+default-features = false
+features = ["std", "perf"]
+version = "1.6.0"
