+++ rust/types/src/lib.rs
@@ -1,0 +1,9 @@
+#![warn(rust_2018_idioms, missing_debug_implementations)]
+//! MediaSoup Rust Types
+//!
+//! This library provides Rust types for interacting with [mediasoup](https://docs.rs/mediasoup) crate.
+//!
+
+pub mod data_structures;
+pub mod rtp_parameters;
+pub mod scalability_modes;
